Update README.md
This project consists of 2 Python programs. One allows the user to input a set of data points, and the range of slopes and intercepts they want to test. The program calculates the error of each slope, intercept pair using the sum of squared errors (SSE) method. It then calculates the optimal regression line and plots it along with a 3D graph that displays all possible slope, intercept combinations and their error. The second program takes a set of data points provided by the user and calculates and plots a curve fitted to the data points. This program does not use built in functions so that the algorithm used for curve fitting is more evident.